% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_query.R
\name{all_tables}
\alias{all_tables}
\title{Lists all tables within DB}
\usage{
all_tables(conn = last_connection())
}
\arguments{
\item{conn}{PqConnection. The database connection to use. By default
this is the last connection opened with this package, or if non-existent,
a new connection using \code{getOption('database.yml')}.}
}
\description{
Lists all tables within DB
}
\seealso{
\code{\link{postgresql_connection}}
}

