% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/berdie-package.R
\docType{package}
\name{berdie}
\alias{berdie}
\alias{berdie-package}
\title{Berdie is an R package for synchronizing data sources across two separate layers.}
\description{
Berdie is an R package for synchronizing data sources across two separate layers.
}

