% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/csv_from_sql.R
\name{csv_from_sql}
\alias{csv_from_sql}
\title{Creates a .csv file from a text SQL query or .SQL file}
\usage{
csv_from_sql(sql, output_csv, use.pp = TRUE, ...)
}
\arguments{
\item{sql}{character.SQL filepath or SQL Query.}

\item{output_csv}{character. The filepath of the query's results.  (Include .csv for best results)}

\item{use.pp}{logical. Utilizes the pp() function to run R code contained within hastag brackets #{R_Code}}
}
\description{
Creates a .csv file from a text SQL query or .SQL file
}
\examples{
\dontrun{
csv_from_sql("/path/to/query.SQL", "/path/to/file.csv")
}
}
\author{
Mike Bruno
}

