% Generated by roxygen2 (4.0.0): do not edit by hand
\name{postgresql_connection}
\alias{postgresql_connection}
\title{Obtain a connection to a PostGreSQL database.}
\usage{
postgresql_connection(database.yml, verbose = TRUE, strict = TRUE)
}
\arguments{
\item{database.yml}{character. The location of the database.yml file
to use. This could, for example, be directory from a Rails project.}

\item{verbose}{logical. Whether or not to print messages indicating
loading is in progress.}
}
\value{
JDBCConnection representing the activated PostGreSQL connection.
}
\description{
By current default, the database configuration will always be read
from the development environment.
}
\details{
Your database.yml should look like:

development:
  adapter: postgresql
  database: <name>
  host: <domain>
  port: <port #>
  username: <username>
  password: <password>
}

