% Generated by roxygen2 (4.0.0): do not edit by hand
\name{run_query}
\alias{run_query}
\title{Run an SQL query and return the results as a data.frame}
\usage{
run_query(query, conn = last_connection())
}
\arguments{
\item{query}{character. The SQL query to run.}

\item{conn}{JDBCConnection. The database connection to use. By default
this is the last connection opened with this package, or if non-existent,
a new connection using \code{getOption('database.yml')}.}
}
\value{
a data.frame with the results of the query
}
\description{
Run an SQL query and return the results as a data.frame
}
\seealso{
\code{\link{postgresql_connection}}
}

