% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{berdie}
\alias{berdie}
\alias{berdie-package}
\title{Berdie is an R package for synchronizing data sources across two separate layers.}
\description{
Berdie is an R package for synchronizing data sources across two separate layers.
}

