% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postgresql_connection.R
\name{postgresql_connection}
\alias{postgresql_connection}
\title{Obtain a connection to a PostGreSQL database.}
\usage{
postgresql_connection(database.yml, env = "development", verbose = TRUE,
  strict = TRUE)
}
\arguments{
\item{database.yml}{character. The location of the database.yml file
to use. This could, for example, be directory from a Rails project.}

\item{env}{character. What environment to use. The default is
`"development"`.}

\item{verbose}{logical. Whether or not to print messages indicating
loading is in progress.}
}
\value{
PqConnection representing the activated PostGreSQL connection.
}
\description{
By default, this function will read from the `development` environment.
}
\details{
Your database.yml should look like:

development:
  adapter: postgresql
  database: <name>
  host: <domain>
  port: <port #>
  username: <username>
  password: <password>
}

