% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prep_sql}
\alias{prep_sql}
\title{Converts SQL files into a format readable by the run_query() function.}
\usage{
prep_sql(path)
}
\arguments{
\item{path}{character.  The filepath of the SQL file to be formatted}
}
\value{
string of SQL query with '--' comments removed and spacing reduced
}
\description{
Converts SQL files into a format readable by the run_query() function.
}

