% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_query.R
\name{run_query}
\alias{run_query}
\title{Run an SQL query and return the results as a data.frame}
\usage{
run_query(query, conn = last_connection())
}
\arguments{
\item{query}{character. The SQL query to run. This can also be left
blank to obtain the latest database connection.}

\item{conn}{JDBCConnection. The database connection to use. By default
this is the last connection opened with this package, or if non-existent,
a new connection using \code{getOption('database.yml')}.}
}
\value{
a data.frame with the results of the query. If the \code{query}
  parameter was blank, the last database connection will be returned.
  (If no database connection was used, note this may spin up a new one
  dependending on your configuration. See \code{\link{last_connection}}.)
}
\description{
Run an SQL query and return the results as a data.frame
}
\seealso{
\code{\link{postgresql_connection}}
}

